Functionallity for forward and backward movement
