+++ tango.py
@@ -97,21 +97,3 @@ class Tango:
             if pos < self.CENTER + self.TO_EXT:
                 self.tango.bendTurn(port, pos + inc)
                 
-class InputHandler:
-    def __init__(self):
-        os.system('xset r off')
-        self.tango = Tango() # Tango instance to issue commands
-        self.root = tkinter.Tk()
-
-    # Determine what to move from the key pressed.
-    def input_pressed(key):
-        print("Pressed {}".format(key))
-        self.tango.drive(-1)
-
-    # Slow down the robot on key release
-    def input_released(key):
-        print("Released {}".format(key))
-    root.bind('<KeyPress>', method_name)
-    root.bind('<KeyRelease>', method_name)
-
-start = InputHandler()
